refactor: move study types to sharedpackage
