+++ apps/submission/package.json
@@ -27,6 +27,7 @@
 		"start:prod": "pnpm migrate:db:prod && pnpm migrate:pcgl:db:prod && NODE_ENV=production node --import=./register.js src/index.js"
 	},
 	"dependencies": {
+		"@clinical-submission/data-model": "workspace:^",
 		"@clinical-submission/validation": "workspace:^",
 		"@overture-stack/lyric": "^0.17.0",
 		"axios": "^1.10.0",
