+++ packages/data-model/package.json
@@ -1,0 +1,20 @@
+{
+	"name": "@clinical-submission/data-model",
+	"private": true,
+	"version": "0.0.0",
+	"description": "PCGL Submission Data Model",
+	"license": "ISC",
+	"exports": "./dist/index.js",
+	"types": "dist/index.d.ts",
+	"type": "module",
+	"scripts": {
+		"build": "tsc",
+		"prebuild": "rimraf ./dist"
+	},
+	"dependencies": {
+		"zod": "^3.24.2"
+	},
+	"devDependencies": {
+		"typescript": "^5.5.3"
+	}
+}
